backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pymongo==4.7.3
python-dotenv==1.0.1
langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.14
sentence-transformers==3.0.1
